Statefull App - List is stored in app state
- moved render to render.js
- turned App into React class
TODO : load each todo to state using Redux
